# src/Nox.ClientApi.Tests/ServiceMetadata/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nox.ClientApi.Tests",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Countries": {
      "get": {
        "tags": [
          "Countries"
        ],
        "operationId": "api/Countries",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json;odata.metadata=minimal;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=minimal": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=full": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.metadata=none": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.streaming=true": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json;odata.streaming=false": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "application/octet-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Countries"
        ],
        "operationId": "api/Countries",
        "requestBody": {
          "content": {
            "application/json;odata.metadata=minimal;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=minimal": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=full": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.metadata=none": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CountryCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json;odata.metadata=minimal;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=minimal": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=full": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.metadata=none": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.streaming=true": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json;odata.streaming=false": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountryBarCodeCreateDto": {
        "required": [
          "barCodeName"
        ],
        "type": "object",
        "properties": {
          "barCodeName": {
            "minLength": 1,
            "type": "string"
          },
          "barCodeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CountryBarCodeDto": {
        "type": "object",
        "properties": {
          "barCodeName": {
            "type": "string",
            "nullable": true
          },
          "barCodeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CountryCreateDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "population": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "countryDebt": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "firstLanguageCode": {
            "type": "string",
            "nullable": true
          },
          "shortDescription": {
            "type": "string",
            "nullable": true
          },
          "countryShortNames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountryLocalNameCreateDto"
            },
            "nullable": true
          },
          "countryBarCode": {
            "$ref": "#/components/schemas/CountryBarCodeCreateDto"
          }
        },
        "additionalProperties": false
      },
      "CountryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "population": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "countryDebt": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "firstLanguageCode": {
            "type": "string",
            "nullable": true
          },
          "shortDescription": {
            "type": "string",
            "nullable": true
          },
          "physicalWorkplaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkplaceDto"
            },
            "nullable": true
          },
          "countryShortNames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountryLocalNameDto"
            },
            "nullable": true
          },
          "countryBarCode": {
            "$ref": "#/components/schemas/CountryBarCodeDto"
          },
          "deletedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "@odata.etag": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "CountryLocalNameCreateDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "nativeName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CountryLocalNameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "nativeName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CurrencyCode": {
        "enum": [
          8,
          12,
          32,
          36,
          44,
          48,
          50,
          51,
          52,
          60,
          64,
          68,
          72,
          84,
          90,
          96,
          104,
          108,
          116,
          124,
          132,
          136,
          144,
          152,
          156,
          170,
          174,
          188,
          192,
          203,
          208,
          214,
          222,
          230,
          232,
          238,
          242,
          262,
          270,
          292,
          320,
          324,
          328,
          332,
          340,
          344,
          348,
          352,
          356,
          360,
          364,
          368,
          376,
          388,
          392,
          398,
          400,
          404,
          408,
          410,
          414,
          417,
          418,
          422,
          426,
          430,
          434,
          446,
          454,
          458,
          462,
          480,
          484,
          496,
          498,
          504,
          512,
          516,
          524,
          532,
          533,
          548,
          554,
          558,
          566,
          578,
          586,
          590,
          598,
          600,
          604,
          608,
          634,
          643,
          646,
          654,
          682,
          690,
          694,
          702,
          704,
          706,
          710,
          728,
          748,
          752,
          756,
          760,
          764,
          776,
          780,
          784,
          788,
          800,
          807,
          818,
          826,
          834,
          840,
          858,
          860,
          882,
          886,
          901,
          925,
          926,
          927,
          928,
          929,
          930,
          931,
          932,
          933,
          934,
          936,
          938,
          940,
          941,
          943,
          944,
          946,
          947,
          948,
          949,
          950,
          951,
          952,
          953,
          955,
          956,
          957,
          958,
          959,
          960,
          961,
          962,
          963,
          964,
          965,
          967,
          968,
          969,
          970,
          971,
          972,
          973,
          975,
          976,
          977,
          978,
          979,
          980,
          981,
          984,
          985,
          986,
          990,
          994,
          997,
          999
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "MoneyDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currencyCode": {
            "$ref": "#/components/schemas/CurrencyCode"
          }
        },
        "additionalProperties": false
      },
...
      "WorkplaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "greeting": {
            "type": "string",
            "nullable": true
          },
          "belongsToCountryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "belongsToCountry": {
            "$ref": "#/components/schemas/CountryDto"
          },
          "@odata.etag": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}